update: 更新文件 bind_context.cpp

+++ frameworks/js/napi/socket/async_context/src/bind_context.cpp
@@ -77,7 +77,7 @@ int BindContext::GetSocketFd() const
     if (sharedManager_ == nullptr) {
         return -1;
     }
-    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<intptr_t>(sharedManager_->GetData())) : -1;
+    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<uint64_t>(sharedManager_->GetData())) : -1;
 }
 
 bool BindContext::CheckParamsType(napi_value *params, size_t paramsCount)
